Go1.16 1) fix plugin buildmode support with:   a) add !config.ctxt.Flag_shared to MOV* ssa, which don't combine   reloc hi, reloc lo, mov reg; to reloc hi, mov reg(lo)   b) change rewriteGot func to matching a) modify   c) add supported mod with linux/sw64 ports 2) Experimental: enable internal cgo linkmode with linux/sw64 ports   a) change hash entry size to 8bytes, which matching sw64 ABIs   b) change gotplt table to got, which matching sw64 ABIs   c) add DT_LOPROC to plt table   d) change LITERAL/GPRELxx size from 2 to 4, martching go link achivement 3) Experimental: linker change   a) add common relocation suppport in sw64.adddynrel, include pie support   b) rewrite sw64.setupplt, which using unwritable plt   c) change GPDISP and GPRELxx to use new way to reloc targets   d) add GOTTPREL compute at internal linkmode   e) add addpltsym to create plt entry for call

+++ src/cmd/link/internal/ld/data.go
@@ -533,8 +533,7 @@ func (st *relocSymState) relocsym(s loader.Sym, P []byte) {
 				Hi := uint16(o >> 16)
 				Lo := uint16(o)
 				target.Arch.ByteOrder.PutUint16(P[off:], Hi)
-				target.Arch.ByteOrder.PutUint16(P[off+4:], Lo)
-
+				target.Arch.ByteOrder.PutUint16(P[int64(off)+4:], Lo)
 			} else {
 				target.Arch.ByteOrder.PutUint32(P[off:], uint32(o))
 			}
